refactor: implement gemini recommendations

+++ src/components/clients/insurance-needs/advanced-income-replacement-card.tsx
@@ -251,7 +251,7 @@ export function AdvancedIncomeReplacementCard({
                       setExpenseReductionPercent(
                         Math.max(
                           0,
-                          Math.min(100, parseInt(e.target.value) || 0),
+                          Math.min(100, parseFloat(e.target.value) || 0),
                         ),
                       )
                     }
